fix(Formato de preços): Fixa o formato dos preços para passar nos requisitos do teste.

+++ back-end/src/database/models/sale.js
@@ -14,7 +14,7 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.INTEGER,
       foreingKey: true,
     },
-    totalPrice: DataTypes.FLOAT,
+    totalPrice: DataTypes.DECIMAL(4,2),
     deliveryAddress: DataTypes.STRING(100),
     deliveryNumber: DataTypes.STRING(50),
     saleDate: DataTypes.DATE,
